requirements.txt
# ColdCraft - Outreach Engine
flask==3.1.0
python-dotenv==1.1.0
google-generativeai==0.8.5
google-api-python-client==2.165.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
beautifulsoup4==4.13.3
requests==2.32.3
PyMuPDF==1.25.4
python-whois==0.9.5
lxml==5.3.1
